WAITP-1253 Move SessionSwitchingAuthHandler to server-boilerplate Allow it to run prior to session being attached to request
