+++ packages/api-client/src/auth/index.ts
@@ -5,5 +5,4 @@
 
 export * from './BasicAuthHandler';
 export * from './ForwardingAuthHandler';
-export * from './SessionSwitchingAuthHandler';
 export * from './TokenAuthHandler';
